+++ module_main/src/main/java/com/anju/yyk/main/ui/frg/register/RegisterFrg.java
@@ -2,9 +2,11 @@ package com.anju.yyk.main.ui.frg.register;
 
 import android.Manifest;
 import android.app.Activity;
+import android.app.AlertDialog;
 import android.content.Context;
 import android.content.Intent;
 import android.content.pm.PackageManager;
+import android.database.Cursor;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.net.Uri;
@@ -15,7 +17,9 @@ import android.os.Message;
 import android.provider.MediaStore;
 import android.text.TextUtils;
 import android.util.Log;
+import android.view.LayoutInflater;
 import android.view.View;
+import android.widget.Button;
 import android.widget.EditText;
 
 import androidx.core.content.ContextCompat;
@@ -84,6 +88,7 @@ public class RegisterFrg extends BaseMvpFragment<RegisterPresenter, RegisterMode
     private final int CREATED_PHOTO_COMPLETE = 1011;
     /** 准备照片数据*/
     private final int PREPARE_PHOTO_DATA = 1012;
+    private final int ALBUM_RESULT_CODE = 1013;
 
     @BindView(R2.id.recyclerView)
     RecyclerView mRecyclerView;
@@ -118,6 +123,7 @@ public class RegisterFrg extends BaseMvpFragment<RegisterPresenter, RegisterMode
     private List<String> imageNameList = new ArrayList<String>(); // 存放上传成功的图片的路径
     private List<String> failedImageList = new ArrayList<String>(); // 存放上传失败的图片路径
     private String content;
+    private int btnType = -1;
 
     @Override
     public int getLayoutId() {
@@ -147,15 +153,25 @@ public class RegisterFrg extends BaseMvpFragment<RegisterPresenter, RegisterMode
                 .subscribe(new Consumer<Permission>() {
                     @Override
                     public void accept(Permission permission) throws Exception {
+                        if (btnType == 1) {
+                            if (hasPermission(Manifest.permission.WRITE_EXTERNAL_STORAGE)) {
+                                openGallery();
+                            }
+                        } else if (btnType == 2) {
+                            if (hasPermission(Manifest.permission.WRITE_EXTERNAL_STORAGE)
+                                    && hasPermission(Manifest.permission.CAMERA)) {
+                                openCamera();
+                            }
+                        }
                         if (permission.granted) {
-                            if (Manifest.permission.WRITE_EXTERNAL_STORAGE.equals(permission.name)
+                            /*if (Manifest.permission.WRITE_EXTERNAL_STORAGE.equals(permission.name)
                                     && hasPermission(Manifest.permission.CAMERA)) {
                                 openCamera();
                             }
                             if (Manifest.permission.CAMERA.equals(permission.name)
                                     && hasPermission(Manifest.permission.WRITE_EXTERNAL_STORAGE)) {
                                 openCamera();
-                            }
+                            }*/
                             // 用户已经同意该权限
                         } else if (permission.shouldShowRequestPermissionRationale) {
                             // 用户拒绝了该权限，没有选中『不再询问』（Never ask again）,那么下次再次启动时。还会提示请求权限的对话框
@@ -173,16 +189,62 @@ public class RegisterFrg extends BaseMvpFragment<RegisterPresenter, RegisterMode
             public void onItemClick(BaseQuickAdapter adapter, View view, int position) {
                 PhotoEntity entity = photos.get(position);
                 if (entity.getItemType() == PhotoEntity.ADD_TYPE){
-                    if (!hasPermission(Manifest.permission.WRITE_EXTERNAL_STORAGE)) {
+                    showChooseDialog();
+                    /*if (!hasPermission(Manifest.permission.WRITE_EXTERNAL_STORAGE)) {
                         requestPermissions(Manifest.permission.WRITE_EXTERNAL_STORAGE);
                     } else if (!hasPermission(Manifest.permission.CAMERA)) {
                         requestPermissions(Manifest.permission.CAMERA);
                     } else {
                         openCamera();
-                    }
+                    }*/
                 }
             }
         });
+    }
+
+    private void openGallery() {
+        if (!hasPermission(Manifest.permission.WRITE_EXTERNAL_STORAGE)) {
+            requestPermissions(Manifest.permission.WRITE_EXTERNAL_STORAGE);
+        } else {
+            Intent albumIntent = new Intent(Intent.ACTION_PICK);
+            albumIntent.setDataAndType(MediaStore.Images.Media.EXTERNAL_CONTENT_URI, "image/*");
+            startActivityForResult(albumIntent, ALBUM_RESULT_CODE);
+        }
+    }
+
+    private void showChooseDialog() {
+        if (mActivity == null) {
+            return;
+        }
+        AlertDialog.Builder builder = new AlertDialog.Builder(mActivity);
+        View view = LayoutInflater.from(mActivity).inflate(R.layout.home_dialog, null);
+        builder.setView(view);
+        AlertDialog dialog = builder.create();
+        Button chooseBtn = view.findViewById(R.id.choose_btn);
+        Button cameraBtn = view.findViewById(R.id.camera_btn);
+        chooseBtn.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                dialog.dismiss();
+                btnType = 1;
+                openGallery();
+            }
+        });
+        cameraBtn.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                dialog.dismiss();
+                btnType = 2;
+                if (!hasPermission(Manifest.permission.WRITE_EXTERNAL_STORAGE)) {
+                    requestPermissions(Manifest.permission.WRITE_EXTERNAL_STORAGE);
+                } else if (!hasPermission(Manifest.permission.CAMERA)) {
+                    requestPermissions(Manifest.permission.CAMERA);
+                } else {
+                    openCamera();
+                }
+            }
+        });
+        dialog.show();
     }
 
     @Override
@@ -454,7 +516,9 @@ public class RegisterFrg extends BaseMvpFragment<RegisterPresenter, RegisterMode
                                 showToast(R.string.error_photo_bitmap_isnull);
                             }
                         }
-                        file.delete();
+                        if (btnType == 2) {
+                            file.delete();
+                        }
                     }
                     hideLoading();
                     break;
@@ -496,6 +560,31 @@ public class RegisterFrg extends BaseMvpFragment<RegisterPresenter, RegisterMode
                 }
             } else if (resultCode == Activity.RESULT_CANCELED) {
                 showToast(R.string.toast_cancel_take_photo);
+            }
+        } else if (requestCode == ALBUM_RESULT_CODE) {
+            if (resultCode == Activity.RESULT_OK) {
+                if (data != null && data.getData() != null) {
+                    Cursor cursor = null;
+                    try {
+                        String[] pro = {MediaStore.Images.Media.DATA};
+                        cursor = mActivity.getContentResolver().query(data.getData(), pro, null, null, null);
+                        if (cursor == null) {
+                            return;
+                        }
+                        int column_index = cursor.getColumnIndexOrThrow(MediaStore.Images.Media.DATA);
+                        cursor.moveToFirst();
+                        String path = cursor.getString(column_index);
+                        file = new File(path);
+                        if (file.exists() && file.length() > 0) {
+                            handler.sendEmptyMessage(CREATED_PHOTO_COMPLETE);
+                        }
+                    } catch (Exception e) {
+                        e.printStackTrace();
+                    } finally {
+                        if (cursor != null)
+                            cursor.close();
+                    }
+                }
             }
         }
         super.onActivityResult(requestCode, resultCode, data);
